// apps/ui/src/app/client/model/modelDto.ts
/**
 * LMStudio Wrapper API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ModelCapabilitiesDto } from './modelCapabilitiesDto';
import { QuantizationDto } from './quantizationDto';
import { LoadedInstanceDto } from './loadedInstanceDto';


export interface ModelDto { 
    type: ModelDto.TypeEnum;
    publisher: string;
    key: string;
    display_name: string;
    architecture?: object | null;
    quantization: QuantizationDto | null;
    size_bytes: number;
    params_string: object | null;
    loaded_instances: Array<LoadedInstanceDto>;
    max_context_length: number;
    format: ModelDto.FormatEnum | null;
    capabilities?: ModelCapabilitiesDto;
    description?: object | null;
}
export namespace ModelDto {
    export const TypeEnum = {
        Llm: 'llm',
        Embedding: 'embedding'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
    export const FormatEnum = {
        Gguf: 'gguf',
        Mlx: 'mlx'
    } as const;
    export type FormatEnum = typeof FormatEnum[keyof typeof FormatEnum];
}